+++ Firmware/STM32/stm32_spi.cpp
@@ -16,14 +16,14 @@ bool STM32_SPI_t::init(STM32_GPIO_t* sck_gpio, STM32_GPIO_t* miso_gpio, STM32_GP
     rx_dma_ = rx_dma;
 
     if (tx_dma) {
-        if (!tx_dma->init(tx_dmas, DMA_t::MEMORY, DMA_t::PERIPHERAL, DMA_t::ALIGN_8_BIT, DMA_t::LINEAR, DMA_t::LOW)) {
+        if (!tx_dma->init(tx_dmas, DMA_t::MEMORY, DMA_t::PERIPHERAL, DMA_t::ALIGN_8_BIT, DMA_t::LINEAR, DMA_t::LOW, 1)) {
             return false;
         }
         tx_dma->link(hspi, &SPI_HandleTypeDef::hdmatx);
     }
 
     if (rx_dma) {
-        if (!rx_dma->init(rx_dmas, DMA_t::PERIPHERAL, DMA_t::MEMORY, DMA_t::ALIGN_8_BIT, DMA_t::LINEAR, DMA_t::LOW)) {
+        if (!rx_dma->init(rx_dmas, DMA_t::PERIPHERAL, DMA_t::MEMORY, DMA_t::ALIGN_8_BIT, DMA_t::LINEAR, DMA_t::LOW, 1)) {
             return false;
         }
         rx_dma->link(hspi, &SPI_HandleTypeDef::hdmarx);