+++ Firmware/STM32/stm32_usart.cpp
@@ -22,14 +22,14 @@ bool STM32_USART_t::init(uint32_t baudrate, STM32_GPIO_t* tx_gpio, STM32_GPIO_t*
     rx_dma_ = rx_dma;
 
     if (tx_dma) {
-        if (!tx_dma->init(tx_dmas, DMA_t::MEMORY, DMA_t::PERIPHERAL, DMA_t::ALIGN_8_BIT, DMA_t::LINEAR, DMA_t::LOW)) {
+        if (!tx_dma->init(tx_dmas, DMA_t::MEMORY, DMA_t::PERIPHERAL, DMA_t::ALIGN_8_BIT, DMA_t::LINEAR, DMA_t::LOW, 1)) {
             return false;
         }
         tx_dma->link(huart, &UART_HandleTypeDef::hdmatx);
     }
 
     if (rx_dma) {
-        if (!rx_dma->init(rx_dmas, DMA_t::PERIPHERAL, DMA_t::MEMORY, DMA_t::ALIGN_8_BIT, DMA_t::CIRCULAR, DMA_t::LOW)) {
+        if (!rx_dma->init(rx_dmas, DMA_t::PERIPHERAL, DMA_t::MEMORY, DMA_t::ALIGN_8_BIT, DMA_t::CIRCULAR, DMA_t::LOW, 1)) {
             return false;
         }
         rx_dma->link(huart, &UART_HandleTypeDef::hdmarx);